Refactor and CMS fix
